Refactor grid methods
